+++ packages/core/src/managers/useEventManager.ts
@@ -1,6 +1,5 @@
-import { Ref, markRaw } from 'vue';
-
 import type { IUseDragOptions } from '../types';
+import { Ref } from 'vue';
 import { createGlobalState } from '@vueuse/core';
 import { preventEvent } from '../utils/events';
 import { useDnDStore } from '../composables/useDnDStore';
@@ -96,8 +95,7 @@ export const useEventManager = createGlobalState(() => {
 
     (event.target as HTMLElement).blur();
 
-    if (options?.container)
-      activeContainer.component.value = markRaw(options.container);
+    if (options?.container) activeContainer.component.value = options.container;
 
     const { activate, track, deactivate } = useSensor(elementRef, options);