+++ playground/src/App.vue
@@ -8,12 +8,9 @@
   import { useDnDStore } from '@vue-dnd-kit/core';
 
   const routes = computed(() => router.options.routes);
-
-  const store = useDnDStore();
 </script>
 
 <template>
-  <pre>{{ store.hovered }}</pre>
   <div class="playground">
     <header>
       <h1>Vue DnD Kit Playground</h1>